[iris] Consolidate stats schemas and periodic emitters into cluster/stats
Move every iris.* namespace constant and row dataclass into a single
catalog (stats/tables.py) — the wire contract dashboards, alert rules,
and the federation hub key on — absorbing worker/stats.py and the schema
halves of runtime/profile.py and autoscaler/provisioning.py. Producers
stay next to their mechanism and import row types from the catalog;
LogStack resolves every table from it.

Add PeriodicEmitter (stats/emitter.py): the daemon-thread wait-first
loop with log-and-skip failure isolation, and run NodeStatsCollector,
ResourceCollector, PeriodicProfiler, TaskStateCollector, and
AdmissionProber on it, deleting five hand-rolled copies of the same
thread scaffold.
