+++ lib/iris/src/iris/cluster/stats/__init__.py
@@ -1,0 +1,4 @@
+# Copyright The Marin Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Finelog stats: the row schemas every Iris emitter writes and the shared periodic runner."""
