+++ lib/iris/tests/cluster/backends/k8s/test_node_metrics.py
@@ -21,7 +21,7 @@ from iris.cluster.backends.k8s.node_metrics import (
     parse_prometheus,
 )
 from iris.cluster.platforms.k8s.fake import InMemoryK8sService
-from iris.cluster.worker.stats import IrisWorkerStat, WorkerStatus
+from iris.cluster.stats.tables import IrisWorkerStat, WorkerStatus
 from iris.test_util import FakeStatsTable
 
 NODE_EXPORTER_TEXT = """
